Optimized GetTypeIdFromDataType
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@2193 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/add_on/scripthelper/scripthelper.cpp
@@ -691,7 +691,7 @@ int ConfigEngineFromStream(asIScriptEngine *engine, istream &strm, const char *c
 			in::ReplaceSlashQuote(decl);
 
 			// Remove the $ that the engine prefixes the behaviours with
-			int n = decl.find("$");
+			size_t n = decl.find("$");
 			if( n != string::npos )
 				decl[n] = ' ';
